Refactored `ApplicationController` with concerns
